score: Use uintptr_t type for TLS DTV generation number
- Change TLS_Dynamic_thread_vector
generation_number from uint32_t to uintptr_t.

- Using a machine-word type resolves the FIXME concern
about correctness across architectures, since uintptr_t
is guaranteed to match size on both 32-bit and 64-bit
targets.

Updates #4827

+++ cpukit/include/rtems/score/tls.h
@@ -107,10 +107,9 @@ extern const volatile TLS_Configuration _TLS_Configuration;
 
 typedef struct {
   /*
-   * FIXME: Not sure if the generation number type is correct for all
-   * architectures.
-  */
-  uint32_t generation_number;
+   * Use a machine word type for the DTV generation number.
+   */
+  uintptr_t generation_number;
 
   void *tls_blocks[ 1 ];
 } TLS_Dynamic_thread_vector;
